feat(reorg): remove substrate-api and merge into substrate
