+++ codegen/Cargo.toml
@@ -11,8 +11,6 @@ proc-macro2 = "1"
 proc-macro-crate = "1"
 convert_case = "0.6"
 
-substrate_api = { version = "0.1.0", registry = "substrate", path = "../substrate-api" }
-
 [dev-dependencies]
 arcstr = "1"
 serde = { version = "1", features = ["derive"] }
