+++ docs/api/code/prelude.rs.hidden
@@ -7,9 +7,9 @@
 #     CustomLayoutType, InOut, Input, LayoutPort, Node, Output, PortGeometry, ShapePort, Signal,
 # };
 # use substrate::layout::{element::Shape, Cell, HasLayout, HasLayoutImpl, Instance};
+# use substrate::pdk::corner::Corner;
 # use substrate::pdk::layers::LayerId;
 # use substrate::pdk::{Pdk, PdkLayers};
-# use substrate::pdk::corner::Corner;
 # use substrate::supported_pdks;
 # use substrate::{
 #     DerivedLayerFamily, DerivedLayers, Io, Layer, LayerFamily, Layers, LayoutData, LayoutType,
