--- substrate-api/src/lib.rs
@@ -1,11 +1,0 @@
-//! The Substrate analog circuit generator framework.
-#![warn(missing_docs)]
-
-pub mod block;
-pub mod context;
-pub mod error;
-pub mod io;
-pub mod layout;
-pub mod pdk;
-pub mod schematic;
-pub mod simulation;
