+++ substrate/Cargo.toml
@@ -4,10 +4,27 @@ version = "0.1.0"
 edition = "2021"
 
 [dependencies]
+serde = { version = "1", features = ["derive"] }
 arcstr = { version = "1", features = ["serde"] }
+anyhow = "1"
+thiserror = "1"
+once_cell = "1"
+impl-trait-for-tuples = "0.2"
+tracing = "0.1"
+rust_decimal = "1.30"
+rust_decimal_macros = "1.30"
+ena = "0.14"
 duplicate = "1"
 
-geometry = { version = "0.2.0", registry = "substrate", path = "../libs/geometry" }
+cache = { version = "0.0.0", registry = "substrate", path = "../libs/cache" }
 codegen = { version = "0.1.0", registry = "substrate", path = "../codegen" }
-substrate_api = { version = "0.1.0", registry = "substrate", path = "../substrate-api" }
+geometry = { version = "0.2.0", registry = "substrate", path = "../libs/geometry" }
+gds = { version = "0.1.0", registry = "substrate", path = "../gds" }
+opacity = { version = "0.1.0", path = "../libs/opacity", registry = "substrate" }
 scir = { version = "0.1.0", registry = "substrate", path = "../scir" }
+pathtree = { version = "0.1.0", registry = "substrate", path = "../libs/pathtree" }
+uniquify = { version = "0.1.0", registry = "substrate", path = "../libs/uniquify" }
+
+[dev-dependencies]
+toml = "0.7"
+approx = "0.5"
