+++ substrate/src/io/mod.rs
@@ -17,10 +17,12 @@ use serde::{Deserialize, Serialize};
 use tracing::Level;
 
 use crate::{
+    block::Block,
     error::Result,
     layout::error::LayoutError,
     pdk::layers::{HasPin, LayerId},
     schematic::{CellId, HasNestedView, InstanceId, InstancePath},
+    Io,
 };
 
 mod impls;
@@ -482,7 +484,7 @@ impl PortGeometryBuilder {
 /// A simple builder that allows setting data at runtime.
 ///
 /// ```
-/// # use substrate_api::io::OptionBuilder;
+/// # use substrate::io::OptionBuilder;
 /// let mut builder = OptionBuilder::default();
 /// builder.set(5);
 /// assert_eq!(builder.build().unwrap(), 5);
@@ -576,3 +578,42 @@ pub struct ArrayData<T> {
 }
 
 // END TYPES
+
+// BEGIN COMMON IO TYPES
+
+/// The interface to a standard 4-terminal MOSFET.
+#[derive(Debug, Default, Clone, Io)]
+pub struct MosIo {
+    /// The drain.
+    pub d: InOut<Signal>,
+    /// The gate.
+    pub g: Input<Signal>,
+    /// The source.
+    pub s: InOut<Signal>,
+    /// The body.
+    pub b: InOut<Signal>,
+}
+
+/// A trait indicating that a block is a standard 4 terminal MOSFET.
+pub trait Mos: Block<Io = MosIo> {}
+
+impl<T> Mos for T where T: Block<Io = MosIo> {}
+
+/// The interface to which simulation testbenches should conform.
+/// TODO: Add trait bound to ensure testbenches have this IO, need to refactor crates.
+#[derive(Debug, Default, Clone, Io)]
+pub struct TestbenchIo {
+    /// The global ground net.
+    pub vss: InOut<Signal>,
+}
+
+/// The interface for 2-terminal voltage sources.
+#[derive(Debug, Default, Clone, Io)]
+pub struct VsourceIo {
+    /// The positive terminal.
+    pub p: InOut<Signal>,
+    /// The negative terminal.
+    pub n: InOut<Signal>,
+}
+
+// END COMMON IO TYPES
