+++ substrate/src/lib.rs
@@ -1,13 +1,22 @@
 //! The Substrate analog circuit generator framework.
 #![warn(missing_docs)]
+
+extern crate self as substrate;
+
 pub use codegen::*;
 #[doc(inline)]
 pub use geometry;
 #[doc(inline)]
 pub use scir;
-pub use substrate_api::*;
 
-pub mod ios;
+pub mod block;
+pub mod context;
+pub mod error;
+pub mod io;
+pub mod layout;
+pub mod pdk;
+pub mod schematic;
+pub mod simulation;
 
 // Re-exported for procedural macros.
 #[doc(hidden)]
