+++ tests/src/shared/inverter/tb.rs
@@ -10,8 +10,8 @@ use spectre::blocks::{Pulse, Vsource};
 use spectre::{Options, Spectre, Tran};
 use substrate::block::Block;
 use substrate::context::Context;
+use substrate::io::TestbenchIo;
 use substrate::io::{Node, Signal};
-use substrate::ios::TestbenchIo;
 use substrate::pdk::corner::{InstallCorner, Pvt};
 use substrate::schematic::{Cell, HasSchematic};
 use substrate::simulation::data::HasNodeData;
